requirement.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
braceexpand==0.1.7
cachetools==5.5.0
contourpy==1.1.1
cycler==0.12.1
dgl==2.1.0
dill==0.3.9
einops==0.8.1
fairscale==0.4.13
filelock
fsspec==2024.10.0
ftfy==6.2.3
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.67.1
h5py==3.11.0
huggingface-hub==0.26.2
imageio==2.35.1
iopath==0.1.10
Jinja2
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.7.5
mkl-fft
mkl-random
mkl-service==2.4.0
networkx
nltk==3.4.5
numpy
omegaconf==2.3.0
openai==1.86.0
opencv-python==4.11.0.86
packaging==24.1
pandas==2.0.3
pillow
portalocker==3.0.0
protobuf==3.20.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.1.4
PyWavelets==1.4.1
PyYAML
regex==2024.11.6
requests
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.2.0
six==1.16.0
stanfordcorenlp==3.9.1.1
sympy
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tiktoken==0.7.0
timm==1.0.11
tokenizers==0.20.3
torch==2.1.0
torchaudio==2.1.0
torchdata==0.7.0
torchvision==0.16.0
tqdm==4.67.0
transformers==4.46.3
triton==2.1.0
typing_extensions==4.12.2
webdataset==0.2.100
